% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outStamp.R
\name{outStamp}
\alias{outStamp}
\title{Formatted Timestamp for Output Files}
\usage{
outStamp()
}
\description{
Formatted Timestamp for Output Files
}
\examples{
paste0(outStamp(), ".analysis.csv")
}
\keyword{workflow}

