% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stdpks.R
\name{stdpks}
\alias{stdpks}
\title{Load regularly used packages in a single step}
\usage{
stdpks()
}
\arguments{
\item{none}{}
}
\description{
Load regularly used packages in a single step
}
\keyword{packages}
\keyword{workflow}

